Base
Added a simple flake and copied all my Hyprland configuration from
NixDots over to here. Now time to test and make sure it's working
properly and is made in a way that it can be easily used by anyone just
by pulling this repository in their flake by defining it as a dependency
and then adding a couple of extra lines, will try to reduce the number
of lines needed to import and use my Hyprland configuration though trust
me.