Caps Lock Indicator
SwayOSD allows displaying an indicator when the capslock key is pressed
which is handy I suppose if your keyboard doesn't show for whatever
reason and anyways it's nice to not have to look away from the monitor.
I also changed the screenshot commands so the screenshots are saved to
system instead of being copied to the system clipboard. I'm sure I can
do both in one command like save and copy to clipboard but I'm a bit
lazy now to look into it. I've got GCSEs to study for :P