UWSM Not Working
So I dunno what happened but when I start Hyprland with UWSM all I see
is a black screen with only my cursor. The system seems to be running
and responding to everything I do like when I use the keybind to open
the terminal I see my cursor change and when I typed in "systemctl
poweroff" and pressed enter it did power off. But absolutely nothing is
being displayed. Just running "Hyprland" works though so that's odd, so
the problem is something with UWSM I guess. I'm gonna look into it now
starting with reverting changes I made. This was the most recent update
after which this issue started so I removed it.